backend/requirements.txt
flask
flask-cors
sqlalchemy==2.0.21
python-dotenv
pyodbc
pytz
